requirements.txt
click
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
numpy
pandas
Pillow
python-dateutil
pytz
six

torch

torchvision
typing-extension
Werkzeug
